\name{fossilq}
\alias{fossilq}
\docType{data}
\title{
Pollen percentages Aluik Pond
}
\description{
This dataset provides pollen percentages of a core from Aluik Pond, Labrador, Canada (Jordan, 1975).
}
\usage{data(fossilq)}

\format{
 A data frame with 9 rows (samples) and 25 columns (age and taxa).
}

\source{
http://www.lpc.uottawa.ca/data/cpd
}
\references{
Jordan, R.H. 1975. Pollen diagrams from Hamilton Inlet, central Labrador, and their environmental implications for the northern maritime Archaic. Arctic Anthropology 12: 92-116.
}
\examples{
data(fossilq)
dim(fossilq)
}
\keyword{datasets}
