\name{modernq}
\alias{modernq}
\docType{data}
\title{Pollen counts Northern Quebec Treeline}
\description{
This data set contains pollen counts for surface samples of 53 lakes in Northern Quebec, Canada (Gajewski, 1991).
}
\usage{data(modernq)}
\format{
  A data frame with 53 observations (lakes) and 55 variables. The first column contains the elevation of each location, and the other 54 the pollen counts for the taxa found in the samples.
 }


\source{
http://www.lpc.uottawa.ca/data/cpd
}
\references{
Gajewski, K. 1991. Representation pollinique a la limite arbres au Noveau-Quebec. Canadian Journal of Earth Sciences 25: 643-648.
}
\examples{
data(modernq)
dim(modernq)

}
\keyword{datasets}
