\name{paleoMAS-package}
\alias{paleoMAS-package}
\alias{paleoMAS}
\docType{package}
\title{
Paleoecological Analysis
}
\description{
Transfer functions and statistical operations for paleoecology 
}
\details{
\tabular{ll}{
Package: \tab paleoMAS\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2009-12-06\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}

\author{
Alexander Correa-Metrio, Dunia H. Urrego, Kenneth R. Cabrera, Mark B. Bush.

Maintainer: Alexander Correa-Metrio <acorrea@my.fit.edu>
}
\references{
Correa-Metrio, A., M.B. Bush, L.Perez, A. Schwalb, and K. Cabrera. Accepted. Pollen distribution along climatic and biogeographic gradients in northern Central America. The Holocene

Correa-Metrio, A. 2010. Climate and vegetation of the Yucatan Peninsula during the late Pleistocene. PhD Dissertation, Florida Institute of Technology, Melbourne, FL. 194 p.

Correa-Metrio, A., K.R. Cabrera, and M.B. Bush. 2010. Quantifying ecological change through discriminant analysis: a paleoecological example from the Peruvian Amazon. \emph{Journal of Vegetation Science} 21: 695-704.
	
Urrego, D., M.B. Bush, M. Silman, A. Correa-Metrio, M. Ledru, F. Mayle, and B. Valencia. 2009. Millenial-scale ecological changes in tropical South America since the Last Glacial Maximum. In: F. Vimieux, F. Sylvestre, and M. Khodri, editors. Past climate variability from the Last Glacial Maximum to the Holocene in South America and surrounding regions. Springer, Paris. 
}

\keyword{ package }
