\name{paleoMAS-package}
\alias{paleoMAS-package}
\alias{paleoMAS}
\docType{package}
\title{
Paleoecological Analysis
}
\description{
Basic computations for paleoecological data}
\details{
\tabular{ll}{
Package: \tab paleoMAS\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2010-01-16\cr
License: \tab GPL\cr
URL: http://www.research.fit.edu/paleolab/ \cr
LazyLoad: \tab yes\cr
}
}
\author{
Alexander Correa-Metrio, Dunia H. Urrego, Kenneth Cabrera, Mark B. Bush.

Maintainer: Alexander Correa-Metrio <acorrea@my.fit.edu>

}
\references{Correa-Metrio, A., K.R. Cabrera, and M.B. Bush. In review. Deriving millenial-scale moisture availability curves from pollen data: an example from the Peruvian Amazon. Journal of Vegetation Science.
	
Urrego, D., M.B. Bush, M. Silman, A. Correa-Metrio, M. Ledru, F.Mayle, and B. Valencia. 2009. Millenial-scale ecological changes in tropical South America since the Last Glacial Maximum. In: F. Vimieux, F. Sylvestre, and M. Khodri, editors. Past climate variability from the Last Glacial Maximum to the Holocene in South America and sorrounding regions. Springer, Paris. 
}

\keyword{ package }
